+++ src-tauri/src/events/mod.rs
@@ -1,0 +1,1 @@
+pub mod typing;
