+++ src-tauri/src/routes/mod.rs
@@ -1,0 +1,6 @@
+pub mod channels;
+pub mod members;
+pub mod message;
+pub mod servers;
+pub mod session;
+pub mod users;
